Дополнение

+++ src/DS3231.h
@@ -1,13 +1,13 @@
 #ifndef __DS3231_H__
 #define __DS3231_H__
 
-#include <Wire.h>
+#include <stdint.h>
 
 #define DS3231_I2C_ADDRESS  0x68
 
 struct DS3231_datetime_t
 {
-    uint16_t year;      // Год
+    uint8_t year;       // Год
     uint8_t month;      // Месяц
     uint8_t day_month;  // День 1-31
     uint8_t day_week;   // День недели 1=Воскресенье, 7=Суббота
@@ -15,6 +15,7 @@ struct DS3231_datetime_t
     uint8_t minute;     // Минуты
     uint8_t second;     // Секунды
 };
+
 
 
 #endif // __DS3231_H__
